+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 - Updated README.
 - Moved the context origin to 0x0 outside of the draw loop (#67)
 - Renamed BaseWidget to CanvasWidget (#27)
+- Removed autoclip config, as this will be automatic. (#20)
 
 ## 0.1.15